+++ tiny_cnn/layers/max_pooling_layer.h
@@ -104,7 +104,7 @@ public:
     }
 
     virtual const vec_t& back_propagation(const vec_t& current_delta, size_t index) override {
-        const vec_t& prev_out = prev_->output(index);
+        const vec_t& prev_out = prev_->output(static_cast<int>(index));
         const activation::function& prev_h = prev_->activation_function();
         vec_t& prev_delta = prev_delta_[index];
         std::vector<int>& max_idx = out2inmax_[index];
